refactor(#126): 서버 컴포넌트 및 api 예외처리 추가
