+++ src/app/class/create/_components/ClassCreate.tsx
@@ -6,11 +6,11 @@ import { ArrowRightSVG } from '@/icons/svg';
 import { updateClassDraft } from '@/lib/apis/classApi';
 import { useClassCreateStore } from '@/store/classCreate';
 import { classOutputDataProcess } from '@/utils/apiDataProcessor';
-import ClassCategory from './_components/ClassCategory';
-import ClassExplanation from './_components/ClassExplanation';
-import ClassLocation from './_components/ClassLocation';
-import ClassPrice from './_components/ClassPrice';
-import ClassSchedule from './_components/ClassSchedule';
+import ClassCategory from './ClassCategory';
+import ClassExplanation from './ClassExplanation';
+import ClassLocation from './ClassLocation';
+import ClassPrice from './ClassPrice';
+import ClassSchedule from './ClassSchedule';
 import ValidationMessage from '@/components/ValidationMessage/ValidationMessage';
 import { IGetClassDraft, classCreateData } from '@/types/class';
 import { ErrorMessage } from '@/types/types';
